notification_api/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.4
PyJWT==2.6.0
redis==4.5.3
celery==5.3.1
pandas==2.0.1